+++ ZoneBuilder/PhysPresetData.cpp
@@ -9,7 +9,7 @@ void writePhysPreset(zoneInfo_t* info, ZStream* buf, PhysPreset* data)
 	WRITE_STRING(data, sndAliasPrefix);
 }
 
-void* addPhysPreset(zoneInfo_t* info, const char* name, char* data, size_t dataLen)
+void* addPhysPreset(zoneInfo_t* info, const char* name, char* data, int dataLen)
 {
 	if (dataLen > 0) { Com_Error(false, "no custom physPreset files yet\n"); return NULL; }
 	return data;
